+++ DSCResources/DSR_IniSettingsFile/DSR_IniSettingsFile.psm1
@@ -136,6 +136,11 @@ function Set-TargetResource
 
     Assert-ParametersValid @PSBoundParameters
 
+    if (-not (Test-Path -Path $Path))
+    {
+        Out-File -FilePath $Path -Force
+    }
+
     if ($Type -eq 'Secret')
     {
         Write-Verbose -Message ($localizedData.SetIniSettingSecretMessage -f `
@@ -219,6 +224,12 @@ function Test-TargetResource
     )
 
     Assert-ParametersValid @PSBoundParameters
+
+    # Check if file being managed exists. If not return $False.
+    if (-not (Test-Path -Path $Path))
+    {
+        return $false
+    }
 
     if ($Type -eq 'Secret')
     {
@@ -306,11 +317,12 @@ function Assert-ParametersValid
         $Secret
     )
 
-    # Does the file in path exist?
-    if (-not (Test-Path -Path $Path))
+    # Does the file's parent path exist?
+    $parentPath = Split-Path -Path $Path -Parent
+    if (-not (Test-Path -Path $parentPath))
     {
         New-InvalidArgumentException `
-            -Message ($localizedData.FileNotFoundError -f $Path) `
+            -Message ($localizedData.FileParentNotFoundError -f $parentPath) `
             -ArgumentName 'Path'
     } # if
 }