+++ Tests/Unit/DSR_KeyValuePairFile.Tests.ps1
@@ -184,6 +184,56 @@ $($script:testAddedName)=$($script:testText)
 
         #region Function Set-TargetResource
         Describe 'DSR_KeyValuePairFile\Set-TargetResource' {
+            Context 'File does not exist' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Assert-ParametersValid `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -Verifiable
+
+                Mock `
+                    -CommandName Get-Content `
+                    -ParameterFilter { $path -eq $script:testTextFile } `
+                    -MockWith { $null } `
+                    -Verifiable
+
+                Mock `
+                    -CommandName Set-Content `
+                    -ParameterFilter {
+                        ($path -eq $script:testTextFile) -and `
+                        ($value -eq "$script:testName=$script:testText")
+                    } `
+                    -Verifiable
+
+                It 'Should not throw an exception' {
+                    { Set-TargetResource `
+                        -Path $script:testTextFile `
+                        -Name $script:testName `
+                        -Ensure 'Present' `
+                        -Text $script:testText `
+                        -Verbose
+                    } | Should -Not -Throw
+                }
+
+                It 'Should call the expected mocks' {
+                    Assert-VerifiableMock
+                    Assert-MockCalled -CommandName Assert-ParametersValid -Exactly 1
+
+                    Assert-MockCalled `
+                        -CommandName Get-Content `
+                        -ParameterFilter { $path -eq $script:testTextFile } `
+                        -Exactly 1
+
+                    Assert-MockCalled `
+                        -CommandName Set-Content `
+                        -ParameterFilter {
+                            ($path -eq $script:testTextFile) -and `
+                            ($value -eq "$script:testName=$script:testText")
+                        } `
+                        -Exactly 1
+                }
+            }
+
             Context 'File exists and contains matching key that should exist' {
                 # verifiable (should be called) mocks
                 Mock `
@@ -489,6 +539,12 @@ $($script:testAddedName)=$($script:testText)
                     -Verifiable
 
                 Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
+                    -Verifiable
+
+                Mock `
                     -CommandName Get-Content `
                     -ParameterFilter { $path -eq $script:testTextFile } `
                     -MockWith { $script:testFileContent } `
@@ -524,6 +580,12 @@ $($script:testAddedName)=$($script:testText)
                 Mock `
                     -CommandName Assert-ParametersValid `
                     -ModuleName 'DSR_KeyValuePairFile' `
+                    -Verifiable
+
+                Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
                     -Verifiable
 
                 Mock `
@@ -564,6 +626,12 @@ $($script:testAddedName)=$($script:testText)
                     -Verifiable
 
                 Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
+                    -Verifiable
+
+                Mock `
                     -CommandName Get-Content `
                     -ParameterFilter { $path -eq $script:testTextFile } `
                     -MockWith { $script:testFileExpectedTextContent } `
@@ -599,6 +667,12 @@ $($script:testAddedName)=$($script:testText)
                 Mock `
                     -CommandName Assert-ParametersValid `
                     -ModuleName 'DSR_KeyValuePairFile' `
+                    -Verifiable
+
+                Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
                     -Verifiable
 
                 Mock `
@@ -641,6 +715,12 @@ $($script:testAddedName)=$($script:testText)
                     -Verifiable
 
                 Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
+                    -Verifiable
+
+                Mock `
                     -CommandName Get-Content `
                     -ParameterFilter { $path -eq $script:testTextFile } `
                     -MockWith { $script:testFileExpectedSecretContent } `
@@ -677,6 +757,12 @@ $($script:testAddedName)=$($script:testText)
                 Mock `
                     -CommandName Assert-ParametersValid `
                     -ModuleName 'DSR_KeyValuePairFile' `
+                    -Verifiable
+
+                Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
                     -Verifiable
 
                 Mock `
@@ -719,6 +805,12 @@ $($script:testAddedName)=$($script:testText)
                     -Verifiable
 
                 Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
+                    -Verifiable
+
+                Mock `
                     -CommandName Get-Content `
                     -ParameterFilter { $path -eq $script:testTextFile } `
                     -MockWith { $script:testFileExpectedTextContent } `
@@ -757,6 +849,12 @@ $($script:testAddedName)=$($script:testText)
                     -Verifiable
 
                 Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
+                    -Verifiable
+
+                Mock `
                     -CommandName Get-Content `
                     -ParameterFilter { $path -eq $script:testTextFile } `
                     -MockWith { $script:testFileExpectedTextContent } `
@@ -764,12 +862,12 @@ $($script:testAddedName)=$($script:testText)
 
                 It 'Should not throw an exception' {
                     { $script:result = Test-TargetResource `
-                        -Path $script:testTextFile `
-                        -Name $script:testName `
-                        -Ensure 'Present' `
-                        -Text $script:testText.ToUpper() `
-                        -IgnoreValueCase:$true `
-                        -Verbose
+                            -Path $script:testTextFile `
+                            -Name $script:testName `
+                            -Ensure 'Present' `
+                            -Text $script:testText.ToUpper() `
+                            -IgnoreValueCase:$true `
+                            -Verbose
                     } | Should -Not -Throw
                 }
 
@@ -793,6 +891,12 @@ $($script:testAddedName)=$($script:testText)
                 Mock `
                     -CommandName Assert-ParametersValid `
                     -ModuleName 'DSR_KeyValuePairFile' `
+                    -Verifiable
+
+                Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $true } `
                     -Verifiable
 
                 Mock `
@@ -825,13 +929,53 @@ $($script:testAddedName)=$($script:testText)
                         -Exactly 1
                 }
             }
+
+            Context 'File exists and does not contain matching key but it should' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Assert-ParametersValid `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -Verifiable
+
+                Mock `
+                    -CommandName Test-Path `
+                    -ModuleName 'DSR_KeyValuePairFile' `
+                    -MockWith { $false } `
+                    -Verifiable
+
+                It 'Should not throw an exception' {
+                    { $script:result = Test-TargetResource `
+                        -Path $script:testTextFile `
+                        -Name $script:testName.ToUpper() `
+                        -Ensure 'Present' `
+                        -Text $script:testText `
+                        -Verbose
+                    } | Should -Not -Throw
+                }
+
+                It 'Should return false' {
+                    $script:result | Should -Be $false
+                }
+
+                It 'Should call the expected mocks' {
+                    Assert-VerifiableMock
+                    Assert-MockCalled -CommandName Assert-ParametersValid -Exactly 1
+                }
+            }
+
         }
         #endregion
 
         #region Function Assert-ParametersValid
         Describe 'DSR_KeyValuePairFile\Assert-ParametersValid' {
-            Context 'File exists' {
+            Context 'File parent path exists' {
                 # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Split-Path `
+                    -ParameterFilter { $path -eq $script:testTextFile } `
+                    -MockWith { $script:testTextFile } `
+                    -Verifiable
+
                 Mock `
                     -CommandName Test-Path `
                     -ParameterFilter { $path -eq $script:testTextFile } `
@@ -840,9 +984,9 @@ $($script:testAddedName)=$($script:testText)
 
                 It 'Should not throw an exception' {
                     { Assert-ParametersValid `
-                        -Path $script:testTextFile `
-                        -Name $script:testName `
-                        -Verbose
+                            -Path $script:testTextFile `
+                            -Name $script:testName `
+                            -Verbose
                     } | Should -Not -Throw
                 }
 
@@ -852,8 +996,14 @@ $($script:testAddedName)=$($script:testText)
                 }
             }
 
-            Context 'File does not exist' {
+            Context 'File parent path does not exist' {
                 # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Split-Path `
+                    -ParameterFilter { $path -eq $script:testTextFile } `
+                    -MockWith { $script:testTextFile } `
+                    -Verifiable
+
                 Mock `
                     -CommandName Test-Path `
                     -ParameterFilter { $path -eq $script:testTextFile } `
@@ -861,14 +1011,14 @@ $($script:testAddedName)=$($script:testText)
                     -Verifiable
 
                 $errorRecord = Get-InvalidArgumentRecord `
-                    -Message ($localizedData.FileNotFoundError -f $script:testTextFile) `
+                    -Message ($localizedData.FileParentNotFoundError -f $script:testTextFile) `
                     -ArgumentName 'Path'
 
                 It 'Should throw expected exception' {
                     { Assert-ParametersValid `
-                        -Path $script:testTextFile `
-                        -Name $script:testName `
-                        -Verbose
+                            -Path $script:testTextFile `
+                            -Name $script:testName `
+                            -Verbose
                     } | Should -Throw $errorRecord
                 }